[WPT] BFCache: pushState() in BFCache/non-BFCache cases
When doing history navigation to an entry that's created
via pushState, the page should be:

- With BFCache: Restored from BFCache.
- Without BFCache: Loaded from the URL set by `pushState()`.

While this test contradicts the current spec but matches
the desired behavior, and the spec will be
fixed as part of https://github.com/whatwg/html/pull/6315.

Tests pass on Firefox/Safari/Chrome,
except for Safari on non-BFCache case
(just because WebLock doesn't disable BFCache).

Bug: 1107415, 1298336, https://github.com/whatwg/html/issues/6207
Change-Id: I609276fe865fa92409fd7a547777dba222bac36c