Flow threads cannot be subtree layout roots.
LayoutTableCell::UpdateStyleWritingModeFromRow() has special code that
prevents a table cell from becoming a writing-mode root (because the
legacy table implementation cannot handle that). As a workaround it
attempts to set all the children of the table cell as writing mode
roots.

This entire mechanism is somewhat broken (the computed value of
writing-mode on table cells is modified somewhere (probably the style
adjuster), so that if we dynamically insert block children later, they
will inherit the incorrect writing mode), so attempting to fix anything
at all there seems risky.

Instead, just refuse to relayout a flow thread as a subtree root, just
like we already do for table parts (including table cells) and grid
items.

Bug: 1174687
Change-Id: I1df68512965f8f38d56339d8ca6add0c00b76f04